pyproject.toml
[tool.poetry]
name = "mwra-export"
version = "0.1.0"
description = ""
authors = ["Graham Voysey <graham@neurala.com>"]

[tool.poetry.dependencies]
python = "^3.8"
camelot-py = {extras = ["all"], version = "^0.8.2"}
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b1"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mwra-export = "mwra_export:app"

[tool.isort]
skip='venv'
profile='black'

[tool.black]
line-length = 120
target-version = ['py36','py37', 'py38']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.env
    | \venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''